Make the arrows pointing towards the child (usual representation)
Originally, the game was representing the commits arrows pointing as a
timeline (from the parent to the child) probably to mimick a timeline.

As the usual representation of the git tree is the opposite (arrows are
from child to parent, in order for the child to restore easily its
history), this patch display the arrows in the usual way.